requirements.txt
Django==1.10.4
Markdown==2.6.7
gunicorn==19.6.0
PyJWT==1.4.2
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-rest-swagger==2.1.0

